requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.13.1,<3.14
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.16
Pillow==9.3.0
django-cors-headers
requests
django_filter
djangorestframework-simplejwt
djangorestframework-simplejwt[token_blacklist]



